+++ tests/unit/module/Oauth2ModuleTest.php
@@ -200,37 +200,8 @@ class Oauth2ModuleTest extends DatabaseTestCase
     public function createClientProvider()
     {
         return [
-            'Scopes as string' => [
+            'scopes' => [
                 'openid email',
-                null,
-                null,
-            ],
-
-            'Scopes as array of strings' => [
-                ['openid', 'email'],
-                null,
-                null,
-            ],
-
-            'Scopes as array of Oauth2Scopes' => [
-                function() {
-                    return [
-                        new Oauth2Scope(['id' => 1]), // Test based on primary key (identifier: 'openid')
-                        new Oauth2Scope(['identifier' => 'email']),
-                    ];
-                },
-                null,
-                null,
-            ],
-
-            'Scopes as array of arrays' => [ // Key is scope identifier, value is the config for the `client_scope` junction table)
-                [
-                    'openid' => [
-                        'applied_by_default' => 1,
-                        'required_on_authorization' => 1,
-                    ],
-                    'email' => [],
-                ],
                 null,
                 null,
             ],
@@ -247,70 +218,6 @@ class Oauth2ModuleTest extends DatabaseTestCase
                 true,
             ],
         ];
-    }
-
-    public function testCreateClientNonExistingScope()
-    {
-        $this->mockConsoleApplication();
-
-        $this->expectExceptionMessage('No scope with identifier "does-not-exists" found.');
-        $client = Oauth2Module::getInstance()->createClient(
-            'test',
-            'test',
-            Oauth2Module::GRANT_TYPE_AUTH_CODE,
-            'https://localhost/test',
-            Oauth2ClientInterface::TYPE_CONFIDENTIAL,
-            'very_secret',
-            'does-not-exists',
-        );
-    }
-
-    public function testCreateClientInvalidScopesType()
-    {
-        $this->mockConsoleApplication();
-
-        $this->expectExceptionMessage('$scopes must be a string or an array.');
-        Oauth2Module::getInstance()->createClient(
-            'test',
-            'test',
-            Oauth2Module::GRANT_TYPE_AUTH_CODE,
-            'https://localhost/test',
-            Oauth2ClientInterface::TYPE_CONFIDENTIAL,
-            'very_secret',
-            new \stdClass(),
-        );
-    }
-
-    public function testCreateClientInvalidScopesArrayValueType()
-    {
-        $this->mockConsoleApplication();
-
-        $this->expectExceptionMessage('If $scopes is an array, its values must be a string, array or an instance of ' . Oauth2ScopeInterface::class . '.');
-        Oauth2Module::getInstance()->createClient(
-            'test',
-            'test',
-            Oauth2Module::GRANT_TYPE_AUTH_CODE,
-            'https://localhost/test',
-            Oauth2ClientInterface::TYPE_CONFIDENTIAL,
-            'very_secret',
-            [new \stdClass()],
-        );
-    }
-
-    public function testCreateClientInvalidScopesMissingUniqueIdentifier()
-    {
-        $this->mockConsoleApplication();
-
-        $this->expectExceptionMessage('Element 0 in $scope should specify either the scope id or its identifier.');
-        Oauth2Module::getInstance()->createClient(
-            'test',
-            'test',
-            Oauth2Module::GRANT_TYPE_AUTH_CODE,
-            'https://localhost/test',
-            Oauth2ClientInterface::TYPE_CONFIDENTIAL,
-            'very_secret',
-            [new Oauth2Scope()],
-        );
     }
 
     public function testCreateClientInvalidSecret()