+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Please check the [Upgrading Instructions](UPGRADE.md) when upgrading to a newer 
 
 ### Added
 - Added `Oauth2Module::$openIdConnectProviderConfigurationInformationPath` to configure OIDC config info endpoint. (rhertogh) 
+- Additional getters and setters for Client properties. (rhertogh)
+- Added `Oauth2ClientInterface::syncClientScopes()` to add/remove/update the client-scope relation. (rhertogh)
 
 ### Changed
 - The `Oauth2ModelRepositoryInterface` now extends `Oauth2RepositoryInterface` and introduced `findModelByPk($pk)`.  (rhertogh)